move SpellList into spellcasting module
